requirements.txt
fastapi
uvicorn
python-dotenv
crewai
crewai-tools
langchain